Add Group Name to ExportAppUsers
When an application is assigned via multiple groups, a user will may be in some or all of those multiple groups. This change would export the group via which each user is assigned the app.

Note: this should also help identify groups which, are assigned to an app, however, none of their members are actually getting assigned the app -because they belong to a higher priority group. These groups can, therefore, be removed from the app assignment.

(2nd attempt)